backend/requirements.txt
# FastAPI & Web Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.18
python-dotenv==1.0.0
slowapi==0.1.9  # Rate limiting

# Database Drivers
motor==3.3.1 # Async MongoDB
pymongo==4.7.0
redis==5.0.1
pinecone==8.0.0

# ORM & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
PyJWT==2.10.1

# AI/ML & Agentic Frameworks
langgraph==0.0.9
langchain==0.2.5
langchain-community==0.3.27
langchain-groq==0.0.1
sentence-transformers==2.2.2
numpy==1.24.3

# API & HTTP
httpx==0.25.1
requests==2.32.4

# Data & Scraping (Optional for Scrapy)
beautifulsoup4==4.12.2
lxml==4.9.4
selenium==4.15.2

# GitHub Integration
PyGithub==2.1.1

# Utilities
python-json-logger==2.0.7
click==8.1.7
